Wake Lock, Notifications, Meta Data Retriever
Wake lock has been enabled on the media player to ensure the device will
not sleep during playback.
Notifications have been created, currently indicating the Playing and
Paused states. Naturally this now means the service is in the foreground
during playback, reducing the likelihood of the OS destroying the
service.
MediaMetadataRetriever wrapper to allow an easy way for the service to
retrieve a tracks title and artist.