Null reference fix, default track selection
Fixed a null reference exception that occurred when attempting to
retrieve the current track when it has not been set.
On launch we default the current track to last played track. If the last
played track does not exist on the device, we default to the first track
in the list.